% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.ftn_request}
\alias{.ftn_request}
\title{FTN api request}
\usage{
.ftn_request(
  .endpoint,
  ...,
  .base_url = "https://data.ftndata.com",
  .api_key = Sys.getenv("FTN_API_KEY")
)
}
\arguments{
\item{...}{query parameters}

\item{.base_url}{base url, default data.ftndata.com}

\item{.api_key}{api key, default to envvar FTN_API_KEY}

\item{.url}{url request}

\item{.name}{description}
}
\description{
FTN api request
}
\seealso{
\url{https://app.swaggerhub.com/apis-docs/FTN-Data/FTN-NFL-API/1.0.0}
}
