% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflverse.R
\name{ftn_nflverse}
\alias{ftn_nflverse}
\title{Get nflverse data}
\usage{
ftn_nflverse(gid = NULL, year = NULL, week = NULL)
}
\arguments{
\item{gid, year, week}{args to specify games to return}
}
\description{
Retrieves FTN-provided nflverse-specific data
}
\seealso{
\href{https://docs.google.com/document/d/12WEm0O9TG7VQsn--fkCQAE_3735owJxfxZ8SvysceZU/edit}{FTN's data docs}
}
