% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participation.R
\name{ftn_participation}
\alias{ftn_participation}
\title{Load and parse ftn participation data}
\usage{
ftn_participation(season)
}
\arguments{
\item{season}{4 digit season to load participation data for}
}
\value{
A parsed dataframe
}
\description{
The function takes season as argument and loads raw participation data from
a nflverse-ftn release. We get data for completed seasons and release the raw
file.
}
