% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflverseftn-package.R
\docType{package}
\name{nflverseftn-package}
\alias{nflverseftn}
\alias{nflverseftn-package}
\title{nflverseftn: Loads data from FTN Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nflverse/nflverse-ftn}
  \item Report bugs at \url{https://github.com/nflverse/nflverse-ftn/issues}
}

}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID})

Other contributors:
\itemize{
  \item FTN Network Inc [copyright holder]
}

}
\keyword{internal}
